processor.py
import numpy as np
from osgeo import gdal

# Here I'm just doubling the values, but in a real scenario,
# I'd implement more complex ecosystem service calculations
def process_geotiff(input_path, output_path):
  # Open the input GeoTIFF
  ds = gdal.Open(input_path)
  if ds is None:
      raise ValueError("Could not open the input file")

  # Read the data into a numpy array
  band = ds.GetRasterBand(1)
  array = band.ReadAsArray()

  # Process the data
  processed_array = array * 2

  # Create a new GeoTIFF file
  driver = gdal.GetDriverByName("GTiff")
  out_ds = driver.Create(output_path, ds.RasterXSize, ds.RasterYSize, 1, gdal.GDT_Float32)

  # Write the processed data
  out_band = out_ds.GetRasterBand(1)
  out_band.WriteArray(processed_array)

  # Copy the projection and geotransform information
  out_ds.SetProjection(ds.GetProjection())
  out_ds.SetGeoTransform(ds.GetGeoTransform())

  # Close the datasets
  ds = None
  out_ds = None

  return output_path

# Analyze remote sensing measurements and asses whether target 
# contains live green vegetation
def calculate_ndvi(input_path, output_path):
  ds = gdal.Open(input_path)
  
  # Assuming band 4 is NIR and band 3 is Red
  nir = ds.GetRasterBand(4).ReadAsArray().astype(np.float32)
  red = ds.GetRasterBand(3).ReadAsArray().astype(np.float32)
  
  # Avoid division by zero
  np.seterr(divide='ignore', invalid='ignore')
  
  # Calculate NDVI
  ndvi = (nir - red) / (nir + red)
  
  # Create the output image
  driver = gdal.GetDriverByName("GTiff")
  outdata = driver.Create(output_path, ds.RasterXSize, ds.RasterYSize, 1, gdal.GDT_Float32)
  outdata.SetGeoTransform(ds.GetGeoTransform())
  outdata.SetProjection(ds.GetProjection())
  outdata.GetRasterBand(1).WriteArray(ndvi)
  outdata.FlushCache()
  
  return output_path

# Calculate Slope and Aspect topographic factors from a Digital Elevation Model
def calculate_slope_aspect(input_path, output_slope_path, output_aspect_path):
  ds = gdal.Open(input_path)
  elevation = ds.GetRasterBand(1).ReadAsArray().astype(np.float32)
  
  # Calculate pixel size
  gt = ds.GetGeoTransform()
  pixel_size_x = gt[1]
  pixel_size_y = -gt[5]
  
  # Calculate gradients
  dy, dx = np.gradient(elevation, pixel_size_y, pixel_size_x)
  
  # Calculate slope
  slope = np.arctan(np.sqrt(dx*dx + dy*dy)) * 57.29578  # Convert to degrees
  
  # Calculate aspect
  aspect = np.arctan2(-dx, dy) * 57.29578  # Convert to degrees
  aspect = 180 + aspect  # Convert to 0-360 range
  
  # Create output datasets
  driver = gdal.GetDriverByName("GTiff")
  
  # Save slope
  outdata_slope = driver.Create(output_slope_path, ds.RasterXSize, ds.RasterYSize, 1, gdal.GDT_Float32)
  outdata_slope.SetGeoTransform(ds.GetGeoTransform())
  outdata_slope.SetProjection(ds.GetProjection())
  outdata_slope.GetRasterBand(1).WriteArray(slope)
  outdata_slope.FlushCache()
  
  # Save aspect
  outdata_aspect = driver.Create(output_aspect_path, ds.RasterXSize, ds.RasterYSize, 1, gdal.GDT_Float32)
  outdata_aspect.SetGeoTransform(ds.GetGeoTransform())
  outdata_aspect.SetProjection(ds.GetProjection())
  outdata_aspect.GetRasterBand(1).WriteArray(aspect)
  outdata_aspect.FlushCache()
  
  return output_slope_path, output_aspect_path

# Estimate carbon storage based on land cover types
def estimate_carbon_storage(input_path, output_path):
  ds = gdal.Open(input_path)
  land_cover = ds.GetRasterBand(1).ReadAsArray()
  
  # Define carbon storage values for each land cover type (ton C/ha)
  # This is a simplified example. Real values would depend on many factors.
  carbon_values = {
      1: 5,    # Urban
      2: 50,   # Agriculture
      3: 100,  # Forest
      4: 2,    # Barren
      5: 10    # Grassland
  }
  
  # Create a numpy array with the same shape as land_cover
  carbon_storage = np.zeros_like(land_cover, dtype=np.float32)
  
  # Assign carbon storage values based on land cover type
  for lc_type, c_value in carbon_values.items():
      carbon_storage[land_cover == lc_type] = c_value
  
  # Create output dataset
  driver = gdal.GetDriverByName("GTiff")
  outdata = driver.Create(output_path, ds.RasterXSize, ds.RasterYSize, 1, gdal.GDT_Float32)
  outdata.SetGeoTransform(ds.GetGeoTransform())
  outdata.SetProjection(ds.GetProjection())
  outdata.GetRasterBand(1).WriteArray(carbon_storage)
  outdata.FlushCache()
  
  return output_path